Fixed: - auto updating files list after upload - added filter to block direct url on some pages - added filter to access management

+++ src/main/java/Servlets/Remove.java
@@ -1,4 +1,8 @@
-import database.Connector;
+package Servlets;
+
+import database.Remover;
+import org.apache.log4j.Logger;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -6,20 +10,18 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
 import java.io.IOException;
-import java.sql.SQLException;
 
 @WebServlet("/remove")
-public class Remover extends HttpServlet {
+public class Remove extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String path = req.getParameter("path");
-        String sql = "DELETE FROM files where path=\"" + path + '"';
-        try {
-            Connector.execute(sql);
-        } catch (SQLException e) {
-            e.printStackTrace();
+        if (Remover.getInst().remove(path)) {
+            File file = new File("webapps/" + path);
+            file.delete();
+            resp.sendRedirect("/cabinet.jsp");
+        } else {
+            Logger.getLogger("Some error in Remover service");
         }
-        File file = new File("webapps/" + path);
-        file.delete();
     }
 }
